+++ SubCodes/Ptransitionmatrix/CreatePTransitionMatrix.m
@@ -46,7 +46,7 @@ else
             % First, get Gamma
             PolicyProbs=zeros(N_a,2,options.precision,'gpuArray');% PolicyProbs are the corresponding probabilities of each of these two.
 
-            PolicyProbs(:,2)=shiftdim(double(Policy(end-1,:)-1)/(simoptions.ngridinterp+1),1); % probability of upper grid point (end-1 because end is now L2flag)
+            PolicyProbs(:,2)=shiftdim((Policy(end-1,:)-1)/(simoptions.ngridinterp+1),1); % probability of upper grid point (end-1 because end is now L2flag)
             PolicyProbs(:,1)=1-PolicyProbs(:,2); % probability of lower grid point
 
             % Policy_aprime and PolicyProbs are currently [N_a,2]
@@ -101,7 +101,7 @@ else
             % First, get Gamma
             PolicyProbs=zeros(N_a,N_e,2,options.precision,'gpuArray');% PolicyProbs are the corresponding probabilities of each of these two.
 
-            PolicyProbs(:,:,2)=shiftdim(double(Policy(end-1,:,:)-1)/(simoptions.ngridinterp+1),1); % probability of upper grid point (end-1 because end is now L2flag)
+            PolicyProbs(:,:,2)=shiftdim((Policy(end-1,:,:)-1)/(simoptions.ngridinterp+1),1); % probability of upper grid point (end-1 because end is now L2flag)
             PolicyProbs(:,:,1)=1-PolicyProbs(:,:,2); % probability of lower grid point
 
             % Policy_aprime and PolicyProbs are currently [N_a,N_e,2]
@@ -154,7 +154,7 @@ else
             % First, get Gamma
             PolicyProbs=zeros(N_a,N_z,2,options.precision,'gpuArray');% PolicyProbs are the corresponding probabilities of each of these two.
 
-            PolicyProbs(:,:,2)=shiftdim(double(Policy(end-1,:,:)-1)/(simoptions.ngridinterp+1),1); % probability of upper grid point (end-1 because end is now L2flag)
+            PolicyProbs(:,:,2)=shiftdim((Policy(end-1,:,:)-1)/(simoptions.ngridinterp+1),1); % probability of upper grid point (end-1 because end is now L2flag)
             PolicyProbs(:,:,1)=1-PolicyProbs(:,:,2); % probability of lower grid point
 
             % Policy_aprime and PolicyProbs are currently [N_a,N_z,2]
@@ -210,7 +210,7 @@ else
             % First, get Gamma
             PolicyProbs=zeros(N_a,N_z*N_e,2,options.precision,'gpuArray');% PolicyProbs are the corresponding probabilities of each of these two.
 
-            PolicyProbs(:,:,2)=shiftdim(double(Policy(end-1,:,:)-1)/(simoptions.ngridinterp+1),1); % probability of upper grid point (end-1 because end is now L2flag)
+            PolicyProbs(:,:,2)=shiftdim((Policy(end-1,:,:)-1)/(simoptions.ngridinterp+1),1); % probability of upper grid point (end-1 because end is now L2flag)
             PolicyProbs(:,:,1)=1-PolicyProbs(:,:,2); % probability of lower grid point
 
             % Policy_aprime and PolicyProbs are currently [N_a,N_z*N_e,2]
@@ -286,7 +286,7 @@ else
             % First, get Gamma
             PolicyProbs=zeros(N_a,N_semiz,2,options.precision,'gpuArray'); % PolicyProbs are the corresponding probabilities of each of these two.
 
-            PolicyProbs(:,:,2)=shiftdim(double(Policy(end-1,:,:)-1)/(simoptions.ngridinterp+1),1); % probability of upper grid point (end-1 because end is now L2flag)
+            PolicyProbs(:,:,2)=shiftdim((Policy(end-1,:,:)-1)/(simoptions.ngridinterp+1),1); % probability of upper grid point (end-1 because end is now L2flag)
             PolicyProbs(:,:,1)=1-PolicyProbs(:,:,2); % probability of lower grid point
 
             N_semizshort=max(max(max(sum((pi_semiz>0),2))));
@@ -378,7 +378,7 @@ else
             % First, get Gamma
             PolicyProbs=zeros(N_a,N_semiz*N_e,2,options.precision,'gpuArray'); % PolicyProbs are the corresponding probabilities of each of these two.
 
-            PolicyProbs(:,:,2)=shiftdim(double(Policy(end-1,:,:)-1)/(simoptions.ngridinterp+1),1); % probability of upper grid point (end-1 because end is now L2flag)
+            PolicyProbs(:,:,2)=shiftdim((Policy(end-1,:,:)-1)/(simoptions.ngridinterp+1),1); % probability of upper grid point (end-1 because end is now L2flag)
             PolicyProbs(:,:,1)=1-PolicyProbs(:,:,2); % probability of lower grid point
             
             N_semizshort=max(max(max(sum((pi_semiz>0),2))));
@@ -470,7 +470,7 @@ else
             % First, get Gamma
             PolicyProbs=zeros(N_a,N_semiz*N_z,2,options.precision,'gpuArray'); % PolicyProbs are the corresponding probabilities of each of these two.
 
-            PolicyProbs(:,:,2)=shiftdim(double(Policy(end-1,:,:)-1)/(simoptions.ngridinterp+1),1); % probability of upper grid point (end-1 because end is now L2flag)
+            PolicyProbs(:,:,2)=shiftdim((Policy(end-1,:,:)-1)/(simoptions.ngridinterp+1),1); % probability of upper grid point (end-1 because end is now L2flag)
             PolicyProbs(:,:,1)=1-PolicyProbs(:,:,2); % probability of lower grid point
 
             N_semizshort=max(max(max(sum((pi_semiz>0),2))));
@@ -566,7 +566,7 @@ else
             % First, get Gamma
             PolicyProbs=zeros(N_a,N_semiz*N_z*N_e,2,options.precision,'gpuArray'); % PolicyProbs are the corresponding probabilities of each of these two.
 
-            PolicyProbs(:,:,2)=shiftdim(double(Policy(end-1,:,:)-1)/(simoptions.ngridinterp+1),1); % probability of upper grid point (end-1 because end is now L2flag)
+            PolicyProbs(:,:,2)=shiftdim((Policy(end-1,:,:)-1)/(simoptions.ngridinterp+1),1); % probability of upper grid point (end-1 because end is now L2flag)
             PolicyProbs(:,:,1)=1-PolicyProbs(:,:,2); % probability of lower grid point
 
             N_semizshort=max(max(max(sum((pi_semiz>0),2))));
