+++ SubCodes/UnKronPolicyIndexes/UnKronPolicyIndexes5_noz.m
@@ -22,20 +22,20 @@ divisors4=cumprod([1,n_daprime4(1:end-1)])';   % [l_daprime4,1]
 divisors5=cumprod([1,n_daprime5(1:end-1)])';   % [l_daprime5,1]
 
 if vfoptions.gridinterplayer==1
-    Policy=[mod(floor(double(PolicyKron(1,:)-1)./divisors1),n_daprime1(:))+1;
-            mod(floor(double(PolicyKron(2,:)-1)./divisors2),n_daprime2(:))+1;
-            mod(floor(double(PolicyKron(3,:)-1)./divisors3),n_daprime3(:))+1;
-            mod(floor(double(PolicyKron(4,:)-1)./divisors4),n_daprime4(:))+1;
-            mod(floor(double(PolicyKron(5,:)-1)./divisors5),n_daprime5(:))+1;
+    Policy=[mod(floor((PolicyKron(1,:)-1)./divisors1),n_daprime1(:))+1;
+            mod(floor((PolicyKron(2,:)-1)./divisors2),n_daprime2(:))+1;
+            mod(floor((PolicyKron(3,:)-1)./divisors3),n_daprime3(:))+1;
+            mod(floor((PolicyKron(4,:)-1)./divisors4),n_daprime4(:))+1;
+            mod(floor((PolicyKron(5,:)-1)./divisors5),n_daprime5(:))+1;
             PolicyKron(6,:);
             PolicyKron(7,:)];
     Policy=reshape(Policy,[l_daprime1+l_daprime2+l_daprime3+l_daprime4+l_daprime5+2,n_a]);
 else
-    Policy=[mod(floor(double(PolicyKron(1,:)-1)./divisors1),n_daprime1(:))+1;
-            mod(floor(double(PolicyKron(2,:)-1)./divisors2),n_daprime2(:))+1;
-            mod(floor(double(PolicyKron(3,:)-1)./divisors3),n_daprime3(:))+1;
-            mod(floor(double(PolicyKron(4,:)-1)./divisors4),n_daprime4(:))+1;
-            mod(floor(double(PolicyKron(5,:)-1)./divisors5),n_daprime5(:))+1];
+    Policy=[mod(floor((PolicyKron(1,:)-1)./divisors1),n_daprime1(:))+1;
+            mod(floor((PolicyKron(2,:)-1)./divisors2),n_daprime2(:))+1;
+            mod(floor((PolicyKron(3,:)-1)./divisors3),n_daprime3(:))+1;
+            mod(floor((PolicyKron(4,:)-1)./divisors4),n_daprime4(:))+1;
+            mod(floor((PolicyKron(5,:)-1)./divisors5),n_daprime5(:))+1];
     Policy=reshape(Policy,[l_daprime1+l_daprime2+l_daprime3+l_daprime4+l_daprime5,n_a]);
 end
